requirements.txt
torch>=2.0.0
gradio>=3.0.0
numpy>=1.20.0
soundfile>=0.12.1
chatterbox-tts